add scroll handler when dropdown menu is attached
