+++ tests/unit/TypeAhead.spec.js
@@ -46,79 +46,6 @@ describe("Moving the Typeahead Pointer", () => {
     expect(Select.vm.typeAheadPointer).toEqual(2);
   });
 
-  describe("Automatic Scrolling", () => {
-    it("should check if the scroll position needs to be adjusted on up arrow keyDown", () => {
-      const Select = mountDefault();
-      const spy = jest.spyOn(Select.vm, "maybeAdjustScroll");
-
-      Select.vm.typeAheadPointer = 1;
-
-      Select.find({ ref: "search" }).trigger("keyup.up");
-      expect(spy).toHaveBeenCalled();
-    });
-
-    it("should check if the scroll position needs to be adjusted on down arrow keyDown", () => {
-      const Select = mountDefault();
-      const spy = jest.spyOn(Select.vm, "maybeAdjustScroll");
-
-      Select.vm.typeAheadPointer = 1;
-
-      Select.find({ ref: "search" }).trigger("keyup.down");
-      expect(spy).toHaveBeenCalled();
-    });
-
-    /**
-     * This test fails despite working in the browser.
-     * After many attempts to get it to pass, it's been
-     * rewritten below.
-     */
-    it.skip("should check if the scroll position needs to be adjusted when filtered options changes", () => {
-      const Select = mountDefault();
-      const spy = jest.spyOn(Select.vm, "maybeAdjustScroll");
-
-      Select.vm.search = "two";
-
-      expect(spy).toHaveBeenCalled();
-    });
-
-    it("should scroll up if the pointer is above the current viewport bounds", () => {
-      const Select = mountDefault();
-      const spy = jest.spyOn(Select.vm, "scrollTo");
-
-      Select.setMethods({
-        pixelsToPointerTop() {
-          return 1;
-        },
-        viewport() {
-          return { top: 2, bottom: 0 };
-        }
-      });
-
-      Select.vm.maybeAdjustScroll();
-
-      expect(spy).toHaveBeenCalledWith(1);
-    });
-
-    it("should scroll down if the pointer is below the current viewport bounds", () => {
-      const Select = mountDefault();
-      const spy = jest.spyOn(Select.vm, "scrollTo");
-
-      Select.setMethods({
-        pixelsToPointerBottom() {
-          return 2;
-        },
-        viewport() {
-          return { top: 0, bottom: 1 };
-        }
-      });
-
-      Select.vm.maybeAdjustScroll();
-      expect(spy).toHaveBeenCalledWith(
-        Select.vm.viewport().top + Select.vm.pointerHeight()
-      );
-    });
-  });
-
   describe("Measuring pixel distances", () => {
     it("should calculate pointerHeight as the offsetHeight of the pointer element if it exists", () => {
       const Select = mountDefault();
